Fixed a condition that was comparing a String and a Fixnum.
The problem was that in the specs :set_dynos was stubed to return 7 (Fixnum), but in the real implementation it returns "7" (String). So I fixed the specs to return "7", and I also added specs for workers and the 3 different cases for the pluralization of "dynos" and "workers" in the outputted sentence.
Also added a spec for Heroku::Client#set_workers.