Ensure database credentials are not displayed for forks on another app
It is possible to create forks and followers on a different app to where
the parent lives. In this case, the database url cannot be matched up
with a config var.

Prior to this patch, the client would display the full URL, which is
something we are getting away from for security reasons. Instead,
this patch shows hostname/port/database name.