Added Toolbelt release tasks
This is so we no longer have to use the toolbelt repo for CLI releases.
Now that the code is consolidated, a single `rake release` will push out
the tgz, zip, gem and deb packages. The pkg and exe builds are not as
important since they auto-update.

Also refactored release code rake tasks.