Relax the version requirement for MultiJson.
Firstly I am trying to update the dependencies on a Rails app and
bundler is attempting to downgrade the heroku gem due to an update
for `multi_json`.

Secondly it appears that the gem is using `MultiJson.dump` and
`MultiJson.load` exclusively (`MultiJson.encode` is an alias to `dump`).
These methods have been the public API since 1.6.0 with no discussion
that I could find about changing them. Considering the surface area
being used of MultiJson and that it's an incredibly common dependency it
seems unnecessary to lock to to a specific micro version.

Finally this replaces the one use of `MultiJson.encode` with
`MultiJson.dump` which is the more common usage in the code.