Add basic specs for public-url; make backup argument optional
Now, heroku pg:backups public-url with no argument will now generate a
URL for the latest backup, as with pgbackups:url. This also moves the
auxiliary information to stdout so that the stdout output of the
command can be used directly to download the backup (e.g., with curl)
or restore from it.