make ps:scale automatically switch tiers
In general, users shouldn't have to be aware of tiers and just be able
to switch between types no matter what tier the type belongs to.

This makes ps:scale automatically switch between tiers when using
ps:scale to change dyno types.

The one exception to this is that if there are existing dynos the user
isn't transferring to a new type, we don't want to change the tier and
effectively move *all* of their dynos to the lowest level on the new
tier.

This change will check to see if the user is trying to enter a state
like that with some dynos on one tier and some dynos on another and
error out.

Wherever possible the code will simply continue to the API and allow it
to show errors rather than add additional validation on the CLI which
may be too strict.