Fix issues with backup scheduling
Previously, if another config var matches the user-provided alias but
its value does *not* match (e.g., the config var for the database the
user is trying to schedule is a substring of another config var with a
different value), we error out. This fixes that and adds specs around
backup scheduling.