Implement automatic SSH tunneling for HPG
When Bastion config vars are present, use them to open a tunnel to
rewrite the URI of the database to the tunnel.

Do this for `pg:psql`, commands like `pg:kill` (which are similar
non-interactive variants), and `pg:push` and `pg:pull`.  Because the
tunneling requires a process to stay alive, switch from "exec"ing psql
to "spawn"ing it as a subprocess, but *only* in the case where
tunneling is required as a hedge against change regular users using
un-tunneled psql.

Example use of pg:psql, on an app that has the bastion-requiring
database "DATABASE" and the regular database "COPPER":

    $ heroku config
    DATABASE_BASTIONS:             @ref:blooming-slyly-8913:bastions
    DATABASE_BASTION_KEY:          @ref:blooming-slyly-8913:bastion-key
    DATABASE_BASTION_REKEYS_AFTER: @ref:blooming-slyly-8913:bastion-rekeys-after
    DATABASE_URL:                  @ref:blooming-slyly-8913:url
    HEROKU_POSTGRESQL_COPPER_URL:  @ref:reading-surely-4913:url

    $ heroku pg:psql -c "select 'hello world'"
    ---> Connecting to DATABASE_URL
      ?column?
    -------------
     hello world
    (1 row)

    $ heroku pg:psql COPPER -c "select 'hello world'"
    ---> Connecting to HEROKU_POSTGRESQL_COPPER_URL
      ?column?
    -------------
     hello world
    (1 row)

Because of the patch to `#exec_sql`, one-shot commands like `pg:kill`,
`pg:outliers`, etc work too:

    $ heroku pg:kill 1234
    WARNING:  PID 1234 is not a PostgreSQL server process
     pg_cancel_backend
    -------------------
     f
    (1 row)

    $ heroku pg:table-size
     name | size
    ------+------
    (0 rows)

Finally, pg:push and pg:pull work (as apparent to the user) exactly
the same as before:

    $ heroku pg:pull DATABASE test
    pg_dump: reading schemas
    pg_dump: reading user-defined tables
    pg_dump: reading extensions
    [...]

    $ heroku pg:push test DATABASE
    [...]