Always display unjoined org apps
This changes the org app list to always display joined and unjoined apps together in one list for parity with Dashboard. Hiding and segregating unjoined apps has proved to confuse users.

This leaves the --all option in place, but updates the documentation to reflect what is actually does for the personal app list.