Use API to resolve attachments for pg:* when no exact match
This is a minimal change because the current state of the pg:* commands is
quite difficult to work with. I don't have sufficient familiarity with the
intended behaviour of each of these commands to make changes with more clarity,
but I can roughly see how it should work...

That is, most of these commands are still Attachment-based and still
APIv2-based. That means we can't easily move to a V3 lookup across the board
because of assumptions in behaviour around implicit attachments (DATABASE),
access to `config_var` (vs `name`). It leads to some surprising output such as:

    $ bin/heroku pg:info budding-busily-2230 -a bjeanes
     !    Ambiguous identifier; multiple matching attachments found: DATABASE, HEROKU_POSTGRESQL_BROWN.

In reality, those two ambiguous attachments are the same add-on and pg:info
would otherwise show similar information for both. Furthermore, since it's not
an attachment name lookup, the `-a`
shouldn't be required.

Nonetheless, in the common case, look up my multiple forms of add-on and
attachment specifiers should work because we just hit the APIv3 for the
resolution, then use the `name` of the V3 representation to derive the
`config_var` which is used to pluck out the correct record from "all"
attachments, as fetched via APIv2.