Consolidate pg:links output for Attachments+Resources
The pg:info command has been updated to display attachments and
resources in a consolidated way that makes more sense in a Shareable
world.

All this really means is not displaying multiple groups of data, if they
are for the same logical database as well as displaying the Resource
name in addition to Attachment names.

Before:

=== HEROKU_POSTGRESQL_IVORY_URL
resource: walking-slowly-1234

=== HEROKU_POSTGRESQL_RED_URL
resource: walking-slowly-1234

After:

=== HEROKU_POSTGRESQL_RED_URL, HEROKU_POSTGRESQL_IVORY_URL (walking-slowly-1234)
...