pg:pull with connection string not just db name
`pg:pull` should convert a connection string for a localhost target
(e.g. to force a port rather than socket connection) into (i.e. `-h` and
`-p`) arguments to postgres cli tools.

This was the case `pg_dump` and `pg_restore` but not for `createdb`.
This patch adjust the call of `createdb` to use the connection string
rather than just the database name.